Refactor to ES6 Modules
